moved admin app to api directory
