--- dook/admin/models.py
@@ -1,1 +1,0 @@
-# Create your models here.
