+++ dook/news/api/views.py
@@ -7,7 +7,7 @@ from rest_framework.response import Response
 from rest_framework.status import HTTP_201_CREATED, HTTP_400_BAD_REQUEST
 from rest_framework.viewsets import GenericViewSet
 
-from dook.admin.permissions import IsExpert, IsFactChecker
+from dook.api.admin.permissions import IsExpert, IsFactChecker
 from dook.news.api.filters import (
     ExpertNewsFilter,
     FactCheckerNewsFilter,
